Removed libc dependency from objc by using std::os::raw types.

+++ src/declare.rs
@@ -38,7 +38,6 @@ use std::error::Error;
 use std::ffi::CString;
 use std::fmt;
 use std::mem;
-use libc::size_t;
 
 use {Encode, Encoding, Message};
 use runtime::{Class, Imp, NO, Object, Sel, self};
@@ -204,7 +203,7 @@ impl ClassDecl {
     pub fn add_ivar<T>(&mut self, name: &str) where T: Encode {
         let c_name = CString::new(name).unwrap();
         let encoding = CString::new(T::encode().as_str()).unwrap();
-        let size = mem::size_of::<T>() as size_t;
+        let size = mem::size_of::<T>();
         let align = mem::align_of::<T>() as u8;
         let success = unsafe {
             runtime::class_addIvar(self.cls, c_name.as_ptr(), size, align,
